[Delivers #160752374] rename  api.py to view.py and orders to manage.py
